+++ packages/gafl-webapp-service/src/server.js
@@ -35,6 +35,7 @@ import { getPlugins } from './plugins.js'
 import { airbrake } from '@defra-fish/connectors-lib'
 import { addEmptyFragmentToUri, addLanguageCodeToUri } from './processors/uri-helper.js'
 import fs from 'fs'
+import { createRequire } from 'node:module'
 
 airbrake.initialise()
 let server
@@ -127,9 +128,18 @@ const logGtmConfig = gtmContainerId => {
   }
 }
 
+const getGovUKFrontendRootPath = async () => {
+  // we use createRequire as require is not available in ES modules
+  const requireInstance = createRequire(`${process.cwd()}/src/server.js`)
+
+  return path.dirname(requireInstance.resolve('govuk-frontend/package.json'))
+}
+
 const init = async () => {
   await server.register(getPlugins())
   const viewPaths = [...new Set(find.fileSync(/\.njk$/, path.join(Dirname, './src/pages')).map(f => path.dirname(f)))]
+
+  const govukFrontendRootPath = await getGovUKFrontendRootPath()
 
   server.views({
     engines: {
@@ -150,8 +160,8 @@ const init = async () => {
 
     // This needs all absolute paths to work with jest and in normal operation
     path: [
-      path.join(Dirname, 'node_modules', 'govuk-frontend', 'dist', 'govuk'),
-      path.join(Dirname, 'node_modules', 'govuk-frontend', 'dist', 'govuk', 'components'),
+      path.join(govukFrontendRootPath, 'dist', 'govuk'),
+      path.join(govukFrontendRootPath, 'dist', 'govuk', 'components'),
       path.join(Dirname, 'src/pages/layout'),
       path.join(Dirname, 'src/pages/macros'),
       ...viewPaths